fix: api

+++ app/src/main/java/com/example/harrypotter/data/remote/dto/Wand.kt
@@ -2,6 +2,6 @@ package com.example.harrypotter.data.remote.dto
 
 data class Wand(
     val core: String,
-    val length: Int,
+    val length: Double,
     val wood: String
 )
